move add Digit To Cluster from cluster to clusterer
